Added ezWorld::SearchForObject() (#1506)
This is a feature for conveniently searching for objects, either relative to some reference object, or through a global key (and then relative to that).

The search path allows to search both upwards (towards a named parent) and downwards through a series of named children.

Optionally, it also only finds objects with a specific component type.