CreateResource is no longer run under the resource manager lock (#1515)
* CreateResource is no longer run under the resource manager lock

This allows for the resource's `CreateResource` function to load other resources without creating a deadlock.
As the resource creation calls `ezResourceManager::GetOrCreateResource` and `ezResourceManager::CreateResource` are no longer executed in their entirely under the resource manager lock, special care needs to be taken that other threads that acquire the lock in the meantime don't interfere with the resource creation. The most important part is that `ezResourceManager::LoadResourceByType` ignores resources that have the `ezResourceFlags::IsCreatedResource` flag set.

# Misc changes
* Fixed various comments in the `ResourceManager.h`.
* Expose `GetCapacity()` on the `ezIdTable`.
* Fix `ezArrayPtr` lldb custom visualizer.
* Created a base class to `ezRegisterSimpleTestHelper` called `ezRegisterTestHelper`which allow it to be derived from to write custom simple tests like simple renderer tests etc. without needing to rewrite the registration infrastructure.
* Added BEGIN-DOCS-CODE-SNIPPET scopes for docs.