Added a ThirdPersonView component for camera placement (#1533)
The component is used to place a camera behind a character, and have it move in and out depending on the surrounding geometry. It can also look at different positions near the character, depending on whether you look up or down (e.g. over the shoulder).

The component is now used by the Animation sample scene, which was also extended such that the aim IK for the gun now also points up and down along the camera.

Also fixed a bug that made the animation smaple start extremely slowly.