Added support for custom value types to Angel Script arrays (#1544)
* Added support for custom value types to Angel Script arrays

Needed to touch angel script itself for this change: the scriptarray class did only support references or handles to the type for operators - not passing the type by value, which is quite common for operators working on value types.
Also fixed that scene layers were not tracked as dependencies of scenes.

* More tests