Jolt weight categories (#1557)
The Jolt config dialog now allows to set up "weight categories" which are named categories for common weights in a game.
In the dynamic actor component, you can then select the weight through a category, rather than specifying a mass or density.
This makes it easier to tweak weights later, as the number of objects to test can be significantly reduced.
It also makes it easier to create assets, as one can quickly choose a weight from a list.

In the future this will be expanded to also allow for easier tweaking of forces and impulses.