Run clang tidy on self-hosted VM (#1585)
The 2 vcores of free VMs take forever to compile. This moves the clang tidy job to the 8 vcore VM and forces the other jobs to be pinned on the physical machines so the VM is practically only used for clang tidy. This makes it easier to maintain as we don't have to create custom line rendering images for the software reasterizers and should speed up things overall.
The physical machines now have a tag `GPU` and the VMs have the tag `VM` to allow jobs to be properly scheduled.