Picking fix for DX11 (#1593)
* Picking fix for DX11

On DX11 or other platforms that do not support multiple bind groups, we always need to invalidate all bind groups when we switch shaders as another bind group could have overwritten the state of a different one as they all write to the same state in the backend.
E.g. a shader that has only one bind group can overwrite everything. If we then switch to a shader that uses 4 bind groups, while bind groups 1 to 3 have not been touched, they are still dirty as the old 1 BG layout and the new 4 BG layout can overlap outside of BG0.

* Applied code formatting rules

---------

Co-authored-by: ezEngine Automation <ezEngineProject@gmail.com>