Input fix for PTG regression (#1610)
* Input fix for PTG regression

A bit ugly, just s stop gap fix until a proper solution is found. Ideally we would track dependencies between render pipelines what what spawned them and remove those automatically if e.g. a view is destroyed. The entire resource lifetime of extracted views also needs to be revisited.

The basic problem is: If we extract something, it must be rendered the next frame as we can't guarantee the extracted resource lifetime longer than that. So we can't skip frames if rendering is scheduled. On the other hand, updating input events will be lost if there is no view -> world that consumes the changed input state so we can't update input if we are not updating views in a frame. This problem only exists in the editor which schedules rendering without views as it destroys view when closing documents but does not remove the extracted view pipeline for the next frame upon view closure.

* Applied code formatting rules

---------

Co-authored-by: ezEngine Automation <ezEngineProject@gmail.com>