Replaced Quake-style console with an ImGui based console (#1668)
This console is much more user-friendly.
It allows to tweak CVars from a window OR a command widget.
It shows the log messages in a scroll area and allows for filtering.
There are basic stats (time + memory).
The stats and log window can be pinned, which means they also appear as an overlay when the console is closed.
When open, the mouse cursor is shown, and it isn't clipped, so opening the console can be used to click
outside the window.